fix(ai16): preserve immutable peer replay metadata
