+++ docs/adr/ADR-034-minimal-cross-host-https-transport.md
@@ -10,9 +10,9 @@
 ## Decision
 
 Cross-host communication is HTTPS requests to the same daemon HTTP router used
-by Unix UDS and loopback-TCP local clients. A daemon is simultaneously an HTTPS
-listener for allowed peers and an HTTPS client when its post-write router
-selects a remote host.
+by Unix direct-UDS local clients and loopback-TCP local clients. A daemon is
+simultaneously an HTTPS listener for allowed peers and an HTTPS client when its
+post-write router selects a remote host.
 There is no second daemon and no cross-host application service.
 
 The only cross-host transport responsibilities are:
